chore(core): Update submodule to track uncommitted changes
The 'core' submodule is updated to a "dirty" state, signified by the `-dirty` suffix on the commit hash.

This change makes the parent repository track a version of the submodule that contains local modifications which have not yet been committed. that contains local modifications which have not yet been committed.